Add files via upload
See www.rudyrucker.com/capow for information on the Capow project.  Basically it's an art and research program for looking at gnarly, lava-lamp-like cellular automata which are based on linear and non-linear wave equations and reaction-diffusion rules.  Orginally developed for 1D CAs, Capow is now primarily about 2D cellular automata, which create self-generating and real-time-tweakable light shows.  The Capow images are akin to the flickercladding patterns on the robots in Rucker's series of cyberpunk WARE novels.